+++ src/black/mode.py
@@ -263,6 +263,7 @@ class Preview(Enum):
     pyi_blank_line_after_function_docstring = auto()
     hug_comparator = auto()
     parenthesize_tuple_in_yield = auto()
+    normalize_tstring_prefix = auto()
 
 
 UNSTABLE_FEATURES: set[Preview] = {